(feat) Code cleanup in static site
