+++ static-site/gatsby-config.js
@@ -19,16 +19,15 @@ module.exports = {
     {
       resolve: `gatsby-plugin-manifest`,
       options: {
-        name: `gatsby-starter-default`,
-        short_name: `starter`,
+        name: `les-dessins-de-virgile`,
+        short_name: `dessins-virgile`,
         start_url: `/`,
         background_color: `#663399`,
         theme_color: `#663399`,
         display: `minimal-ui`,
-        icon: `src/images/gatsby-icon.png`, // This path is relative to the root of the site.
+        icon: `src/images/favicon.png`, // This path is relative to the root of the site.
       },
     },
-    "load-data-plugin",
     {
       resolve: "gatsby-plugin-react-svg",
       options: {
