+++ static-site/gatsby-node.js
@@ -6,6 +6,29 @@
 
 // You can delete this file if you're not using it
 const path = require("path");
+const {
+  generateArticles,
+  generateCategories,
+  generateLogo,
+  generateAbout,
+} = require("./generate-data-nodes");
+
+exports.sourceNodes = async ({
+  actions,
+  createNodeId,
+  createContentDigest,
+}) => {
+  const { createNode } = actions;
+
+  await Promise.all([
+    generateArticles({ createNode, createNodeId, createContentDigest }),
+    generateCategories({ createNode, createNodeId, createContentDigest }),
+    generateLogo({ createNode, createNodeId, createContentDigest }),
+    generateAbout({ createNode, createNodeId, createContentDigest }),
+  ]);
+
+  return;
+};
 
 exports.onCreateNode = ({ node, actions }) => {
   const { createNodeField } = actions;
