+++ static-site/package.json
@@ -5,6 +5,7 @@
   "version": "0.1.0",
   "author": "Kyle Mathews <mathews.kyle@gmail.com>",
   "dependencies": {
+    "axios": "^0.19.2",
     "classnames": "^2.2.6",
     "gatsby": "^2.21.0",
     "gatsby-image": "^2.4.0",
