--- static-site/plugins/load-data-plugin/package.json
@@ -1,11 +1,0 @@
-{
-  "name": "load-data-plugin",
-  "version": "1.0.0",
-  "main": "index.js",
-  "author": "matthis-d",
-  "license": "MIT",
-  "private": true,
-  "dependencies": {
-    "axios": "^0.19.2"
-  }
-}
