--- static-site/src/components/MaterialButton/MaterialButton.js
@@ -1,21 +1,0 @@
-import React from "react";
-import classNames from "classnames";
-import "./MaterialButton.scss";
-
-const MaterialButton = ({
-  handleClick,
-  text,
-  isActive = false,
-  children = text,
-}) => (
-  <button
-    onClick={handleClick}
-    className={classNames("MaterialButton", {
-      "MaterialButton--active": isActive,
-    })}
-  >
-    {children}
-  </button>
-);
-
-export default MaterialButton;
