+++ static-site/src/components/MaterialLink/MaterialLink.js
@@ -1,0 +1,17 @@
+import React from "react";
+import { Link } from "gatsby";
+
+import "./MaterialLink.scss";
+
+const MaterialLink = ({ to, children, ...props }) => (
+  <Link
+    className="MaterialLink"
+    activeClassName="MaterialLink--active"
+    to={to}
+    {...props}
+  >
+    {children}
+  </Link>
+);
+
+export default MaterialLink;
