+++ static-site/src/components/MaterialLink/MaterialLink.scss
@@ -1,4 +1,4 @@
-.MaterialButton {
+.MaterialLink {
   margin-top: 3rem;
   border: 1px solid rgb(156, 156, 156);
   padding: 0.8rem 2rem;
