+++ static-site/src/components/NavBar/NavBar.scss
@@ -82,7 +82,7 @@
       }
 
       &__Categories {
-        button {
+        a {
           margin: 0.2rem;
           padding: 0.3rem;
         }
