+++ static-site/src/images/favicon.png
