+++ static-site/src/pages/index.js
@@ -1,114 +1,18 @@
-import React, { useState, useEffect } from "react";
+import React from "react";
 import { graphql } from "gatsby";
-import StackGrid from "react-stack-grid";
-import Lightbox from "react-image-lightbox";
 
-import useScreenSize from "../components/useScreenSize";
 import Layout from "../components/Layout/Layout";
+import Articles from "../components/Articles/Articles";
 
-import "../styles/Home.scss";
+import "../styles/app.scss";
 
 const Home = ({ data }) => {
-  const [isOpen, setIsOpen] = useState(false);
-  const [selectedImageIndex, setSelectedImageIndex] = useState(0);
-  const [numberOfItemDisplayed, setNumberOfItemDisplayed] = useState(10);
-  const screenSize = useScreenSize();
-
-  const isMobile = screenSize === "small";
-  const isDesktop = screenSize === "large";
-  const gutterValue =
-    screenSize === "large" ? 30 : screenSize === "medium" ? 15 : 10;
-
-  const articles = data.allArticleType.nodes;
-
-  useEffect(() => {
-    const handleScroll = () => {
-      if (
-        window.innerHeight + window.scrollY >=
-          document.body.offsetHeight - 150 &&
-        articles.length
-      ) {
-        setNumberOfItemDisplayed(numberOfItemDisplayed + 10);
-      }
-    };
-    window.addEventListener("scroll", handleScroll);
-
-    return () => {
-      window.removeEventListener("scroll", handleScroll);
-    };
-  }, [articles, numberOfItemDisplayed]);
-
-  const handleClick = (index) => {
-    setSelectedImageIndex(index);
-    setIsOpen(!isOpen);
-  };
+  const articles = data.allArticleType.edges.map(({ node }) => node);
 
   return (
     <div className="App">
       <Layout>
-        <div className="Home">
-          <StackGrid
-            monitorImagesLoaded={true}
-            gutterWidth={gutterValue}
-            gutterHeight={isMobile ? gutterValue / 2 : gutterValue / 1.2}
-            duration={0}
-            columnWidth={isDesktop ? "30%" : "50%"}
-          >
-            {articles
-              .sort((a, b) =>
-                Math.abs(
-                  new Date(b.creationDate).getTime() -
-                    new Date(a.creationDate).getTime()
-                )
-              )
-              .map((article, index) => {
-                return (
-                  index <= numberOfItemDisplayed && (
-                    <button
-                      className="Home__Article"
-                      key={article.id}
-                      onClick={() => handleClick(index)}
-                    >
-                      <img
-                        className="Home__Article__Image"
-                        src={`http://localhost:4444/medias/${article.url}`}
-                        alt=""
-                      />
-                    </button>
-                  )
-                );
-              })}
-          </StackGrid>
-          {isOpen && (
-            <Lightbox
-              mainSrc={`http://localhost:4444/medias/${articles[selectedImageIndex].url}`}
-              nextSrc={`http://localhost:4444/medias/${
-                articles[
-                  selectedImageIndex !== articles.length - 1
-                    ? selectedImageIndex + 1
-                    : selectedImageIndex
-                ].url
-              }`}
-              prevSrc={`http://localhost:4444/medias/${
-                articles[
-                  selectedImageIndex !== 0
-                    ? selectedImageIndex - 1
-                    : selectedImageIndex
-                ].url
-              }`}
-              enableZoom={false}
-              onCloseRequest={() => setIsOpen(false)}
-              onMovePrevRequest={() =>
-                selectedImageIndex !== 0 &&
-                setSelectedImageIndex(selectedImageIndex - 1)
-              }
-              onMoveNextRequest={() =>
-                selectedImageIndex !== articles.length - 1 &&
-                setSelectedImageIndex(selectedImageIndex + 1)
-              }
-            />
-          )}
-        </div>
+        <Articles articles={articles} />
       </Layout>
     </div>
   );
@@ -117,11 +21,13 @@ const Home = ({ data }) => {
 export const query = graphql`
   query HomePageQuery {
     allArticleType {
-      nodes {
-        id
-        categories
-        url
-        title
+      edges {
+        node {
+          id
+          categories
+          url
+          title
+        }
       }
     }
   }
