+++ static-site/src/styles/About.scss
@@ -6,9 +6,6 @@
   padding: 0 1.5rem;
   min-height: calc(100vh - 10rem);
 
-  &__Title {
-  }
-
   &__Image {
     width: 70%;
   }
